requirements.txt
discord.py
PyQt6